livekit-voice-agent/requirements.txt
livekit-agents>=0.12.16
livekit-plugins-openai>=0.10.17
livekit-plugins-silero>=0.7.4
python-dotenv>=1.0.1
